+++ src/main/java/mocha/Mocha.java
@@ -1,13 +1,8 @@
-import java.io.File;
+package mocha;
+
+import mocha.command.Command;
+
 import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
 
 /**
  * Encapsulates the functionality of a chatbot
