+++ src/main/java/mocha/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package mocha.command;
+
+import mocha.TaskFile;
+import mocha.TaskList;
+import mocha.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, TaskFile storage) {
