+++ src/main/java/mocha/command/DeadlineCommand.java
@@ -1,3 +1,11 @@
+package mocha.command;
+
+import mocha.TaskFile;
+import mocha.TaskList;
+import mocha.Ui;
+import mocha.task.Deadline;
+import mocha.task.Task;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
