+++ src/main/java/mocha/command/DueCommand.java
@@ -1,3 +1,12 @@
+package mocha.command;
+
+import mocha.Parser;
+import mocha.TaskFile;
+import mocha.TaskList;
+import mocha.Ui;
+import mocha.task.Event;
+import mocha.task.Todo;
+
 import java.time.LocalDate;
 
 public class DueCommand extends Command {
