+++ src/main/java/mocha/command/EventCommand.java
@@ -1,3 +1,11 @@
+package mocha.command;
+
+import mocha.TaskFile;
+import mocha.TaskList;
+import mocha.Ui;
+import mocha.task.Event;
+import mocha.task.Task;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
